app/requirements.txt
django
psycopg2
djangorestframework
gevent
gunicorn
celery
redis
django-async-messages-redux
tqdm
django-redis
pillow
django-debug-toolbar
django-querycount
sentry-sdk